+++ packages/pybamm-model-zoo/src/pybamm_model_zoo/linearised_spm/__init__.py
@@ -1,0 +1,6 @@
+from pybamm_model_zoo.linearised_spm.model import (
+    LinearisedOpenCircuitPotential,
+    LinearisedSPM,
+)
+
+__all__ = ["LinearisedOpenCircuitPotential", "LinearisedSPM"]